# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для телефонного справочника",
        "title": "Telephone Book API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/birthday/today": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "birthday"
                ],
                "summary": "Дни рождения сегодня",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Институт",
                        "name": "institute",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.User"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/birthday/tomorrow": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "birthday"
                ],
                "summary": "Дни рождения завтра",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Институт",
                        "name": "institute",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.User"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/departments": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "departments"
                ],
                "summary": "Получить все отделы",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Институт",
                        "name": "institute",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/departments.DepartmentsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "departments"
                ],
                "summary": "Обновить отдел",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Институт",
                        "name": "institute",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Старое название отдела",
                        "name": "department",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Новые данные отдела",
                        "name": "department",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/departments.UpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/departments.UpdateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "departments"
                ],
                "summary": "Создать отдел",
                "parameters": [
                    {
                        "description": "Данные отдела",
                        "name": "department",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/departments.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/departments.CreateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "departments"
                ],
                "summary": "Удалить отдел",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Институт",
                        "name": "institute",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Название отдела",
                        "name": "department",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/departments.DeleteResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/departments/{department}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "departments"
                ],
                "summary": "Получить секции отдела",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Институт",
                        "name": "institute",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Название отдела",
                        "name": "department",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/departments.SectionsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "departments.CreateRequest": {
            "type": "object",
            "properties": {
                "institute": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "sections": {
                    "description": "Optional, can be used to specify sections within the department",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "departments.CreateResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Сообщение об ошибке, если есть. Пример: \"invalid request\"",
                    "type": "string"
                },
                "status": {
                    "description": "Статус ответа: Ok или Error. Пример: \"Ok\"",
                    "type": "string"
                },
                "user_id": {
                    "description": "ID созданного отдела",
                    "type": "integer"
                }
            }
        },
        "departments.DeleteResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Сообщение об ошибке, если есть. Пример: \"invalid request\"",
                    "type": "string"
                },
                "status": {
                    "description": "Статус ответа: Ok или Error. Пример: \"Ok\"",
                    "type": "string"
                }
            }
        },
        "departments.DepartmentsResponse": {
            "type": "object",
            "properties": {
                "departments": {
                    "description": "Список отделов",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Department"
                    }
                },
                "error": {
                    "description": "Сообщение об ошибке, если есть. Пример: \"invalid request\"",
                    "type": "string"
                },
                "status": {
                    "description": "Статус ответа: Ok или Error. Пример: \"Ok\"",
                    "type": "string"
                }
            }
        },
        "departments.SectionsResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Сообщение об ошибке, если есть. Пример: \"invalid request\"",
                    "type": "string"
                },
                "sections": {
                    "description": "Список секций",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Section"
                    }
                },
                "status": {
                    "description": "Статус ответа: Ok или Error. Пример: \"Ok\"",
                    "type": "string"
                }
            }
        },
        "departments.UpdateRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "sections": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "departments.UpdateResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Сообщение об ошибке, если есть. Пример: \"invalid request\"",
                    "type": "string"
                },
                "status": {
                    "description": "Статус ответа: Ok или Error. Пример: \"Ok\"",
                    "type": "string"
                }
            }
        },
...
        "models.Department": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.Section": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "parent_id": {
                    "description": "Optional, can be used to specify parent section",
                    "type": "integer"
                }
            }
        },
...
        "models.User": {
            "type": "object",
            "properties": {
                "birth_date": {
                    "type": "string"
                },
                "cabinet": {
                    "type": "string"
                },
                "department": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "middle_name": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "photo": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "position": {
                    "type": "string"
                },
                "section": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                }
            }
        },
...
        "response.Response": {
            "description": "Базовый ответ API",
            "type": "object",
            "properties": {
                "error": {
                    "description": "Сообщение об ошибке, если есть. Пример: \"invalid request\"",
                    "type": "string"
                },
                "status": {
                    "description": "Статус ответа: Ok или Error. Пример: \"Ok\"",
                    "type": "string"
                }
            }
        },
...
    }
}